pyproject.toml
[project]
name = "trip-planner-agent"
version = "0.1.0"
description = "trip planner multi-agent system using langchain ecosystem"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.8,<2.10",
    "python-dotenv",
    "langgraph",
    "langchain[groq]",
    "amadeus",
    "langchain-community",
    "langchain-google-community[calendar]",
    "langgraph-supervisor",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
